requirements.txt
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
cycler==0.11.0
Flask==2.1.2
fonttools==4.34.2
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.3
MarkupSafe==2.1.1
matplotlib==3.5.2
numpy==1.23.0
opencv-python==4.6.0.66
packaging==21.3
Pillow==9.2.0
pyparsing==3.0.9
python-dateutil==2.8.2
requests==2.28.1
six==1.16.0
torch==1.12.0
torchaudio==0.12.0
torchvision==0.13.0
typing_extensions==4.3.0
urllib3==1.26.9
Werkzeug==2.1.2
wincertstore==0.2
